[FORMAT] Add library-level FileInfo helper usable from pyOpenMS (#9596)
Extracts the inspection logic of the FileInfo TOPP tool into a reusable
OpenMS::FileInfo library class (FORMAT/) so it can be called from C++ and
pyOpenMS, e.g. FileInfo().runAll("data.consensusXML").

- New OpenMS::FileInfo with a structured Result (ranges, counts, per-file-type
  info, statistics, validation, ...) plus toText()/toTSV() renderers.
- TOPPFileInfo now delegates to the library: outputTo_ shrinks to a ~30-line
  wrapper, so the library is the single implementation. The ~20 TOPP_FileInfo_*
  golden CLI tests stay byte-for-byte identical (regression gate).
- nanobind bindings in bind_format.cpp (Result/Options/run/runAll/toText/toTSV
  + Math::SummaryStatistics); adds the stl/tuple caster.
- New class test (FileInfo_test) and pyOpenMS test (test_FileInfo.py).

Implementation detail: the rendering code is moved verbatim into the library
and the produced human-readable/TSV strings are cached on the Result, which
guarantees the CLI output is unchanged. Structured typed fields are populated
alongside for programmatic/pyOpenMS use.
