+++ src/pyOpenMS/bindings/bind_format.cpp
@@ -21,7 +21,9 @@
 #include <OpenMS/FORMAT/FLASHDeconvFeatureFile.h>
 #include <OpenMS/FORMAT/FLASHDeconvSpectrumFile.h>
 #include <OpenMS/FORMAT/FileHandler.h>
+#include <OpenMS/FORMAT/FileInfo.h>
 #include <OpenMS/FORMAT/FileTypes.h>
+#include <OpenMS/MATH/StatisticFunctions.h>
 #include <OpenMS/FORMAT/GNPSMetaValueFile.h>
 #include <OpenMS/FORMAT/GNPSQuantificationFile.h>
 #include <OpenMS/FORMAT/HANDLERS/IndexedMzMLDecoder.h>
@@ -2651,5 +2653,244 @@ or chromatograms only (SRM/MRM) and forwards to the appropriate loader.
     m.def("fromFASTAEntry", [](const OpenMS::FASTAFile::FASTAEntry& fasta) {
         return OpenMS::PEFFEntry::fromFASTAEntry(fasta);
     }, "fasta"_a, "Create a PEFFEntry from a FASTAEntry");
+
+    // -----------------------------------------------------------------------
+    // FileInfo (library-level equivalent of the FileInfo TOPP tool)
+    // -----------------------------------------------------------------------
+    {
+      using OpenMS::FileInfo;
+
+      // SummaryStatistics<vector<double>> (reused for FileInfo statistics / FASTA length stats)
+      nb::class_<OpenMS::Math::SummaryStatistics<std::vector<double>>>(m, "SummaryStatistics",
+          "Summary statistics: count/mean/min/lower-quartile/median/upper-quartile/max/variance")
+          .def(nb::init<>())
+          .def_ro("count", &OpenMS::Math::SummaryStatistics<std::vector<double>>::count)
+          .def_ro("mean", &OpenMS::Math::SummaryStatistics<std::vector<double>>::mean)
+          .def_ro("min", &OpenMS::Math::SummaryStatistics<std::vector<double>>::min)
+          .def_ro("lowerq", &OpenMS::Math::SummaryStatistics<std::vector<double>>::lowerq)
+          .def_ro("median", &OpenMS::Math::SummaryStatistics<std::vector<double>>::median)
+          .def_ro("upperq", &OpenMS::Math::SummaryStatistics<std::vector<double>>::upperq)
+          .def_ro("max", &OpenMS::Math::SummaryStatistics<std::vector<double>>::max)
+          .def_ro("variance", &OpenMS::Math::SummaryStatistics<std::vector<double>>::variance)
+          ;
+
+      auto fi = nb::class_<FileInfo>(m, "FileInfo",
+          "Library-level equivalent of the FileInfo TOPP tool: inspect an OpenMS-readable file "
+          "and return all file-level information as a structured result (also usable from pyOpenMS).");
+
+      nb::class_<FileInfo::Range>(fi, "Range", "[min,max] interval for one dimension")
+          .def(nb::init<>())
+          .def_ro("present", &FileInfo::Range::present)
+          .def_ro("min", &FileInfo::Range::min)
+          .def_ro("max", &FileInfo::Range::max);
+
+      nb::class_<FileInfo::RangeSet>(fi, "RangeSet", "RT / m/z / mobility / intensity ranges")
+          .def(nb::init<>())
+          .def_ro("rt", &FileInfo::RangeSet::rt)
+          .def_ro("mz", &FileInfo::RangeSet::mz)
+          .def_ro("mobility", &FileInfo::RangeSet::mobility)
+          .def_ro("intensity", &FileInfo::RangeSet::intensity)
+          .def_ro("has_mobility", &FileInfo::RangeSet::has_mobility);
+
+      nb::class_<FileInfo::Ranges>(fi, "Ranges", "range categories (MSExperiment carries all four)")
+          .def(nb::init<>())
+          .def_ro("combined", &FileInfo::Ranges::combined)
+          .def_ro("spectra_overall", &FileInfo::Ranges::spectra_overall)
+          .def_ro("per_ms_level", &FileInfo::Ranges::per_ms_level)
+          .def_ro("chromatograms", &FileInfo::Ranges::chromatograms)
+          .def_ro("is_experiment", &FileInfo::Ranges::is_experiment);
+
+      nb::class_<FileInfo::FileMeta>(fi, "FileMeta", "general file header (always populated)")
+          .def(nb::init<>())
+          .def_ro("file_name", &FileInfo::FileMeta::file_name)
+          .def_ro("file_type", &FileInfo::FileMeta::file_type)
+          .def_ro("file_type_name", &FileInfo::FileMeta::file_type_name);
+
+      nb::class_<FileInfo::ExperimentMeta::Contact>(fi, "Contact", "a contact person")
+          .def(nb::init<>())
+          .def_ro("first_name", &FileInfo::ExperimentMeta::Contact::first_name)
+          .def_ro("last_name", &FileInfo::ExperimentMeta::Contact::last_name)
+          .def_ro("email", &FileInfo::ExperimentMeta::Contact::email);
+
+      nb::class_<FileInfo::ExperimentMeta>(fi, "ExperimentMeta", "the -m metadata block (peak files)")
+          .def(nb::init<>())
+          .def_ro("present", &FileInfo::ExperimentMeta::present)
+          .def_ro("document_id", &FileInfo::ExperimentMeta::document_id)
+          .def_ro("date", &FileInfo::ExperimentMeta::date)
+          .def_ro("sample_name", &FileInfo::ExperimentMeta::sample_name)
+          .def_ro("sample_organism", &FileInfo::ExperimentMeta::sample_organism)
+          .def_ro("sample_comment", &FileInfo::ExperimentMeta::sample_comment)
+          .def_ro("instrument_name", &FileInfo::ExperimentMeta::instrument_name)
+          .def_ro("instrument_model", &FileInfo::ExperimentMeta::instrument_model)
+          .def_ro("instrument_vendor", &FileInfo::ExperimentMeta::instrument_vendor)
+          .def_ro("ion_sources", &FileInfo::ExperimentMeta::ion_sources)
+          .def_ro("mass_analyzers", &FileInfo::ExperimentMeta::mass_analyzers)
+          .def_ro("detectors", &FileInfo::ExperimentMeta::detectors)
+          .def_ro("contacts", &FileInfo::ExperimentMeta::contacts);
+
+      nb::class_<FileInfo::ProcessingStep>(fi, "ProcessingStep", "a data-processing step (-p)")
+          .def(nb::init<>())
+          .def_ro("software_name", &FileInfo::ProcessingStep::software_name)
+          .def_ro("software_version", &FileInfo::ProcessingStep::software_version)
+          .def_ro("completion_time", &FileInfo::ProcessingStep::completion_time)
+          .def_ro("actions", &FileInfo::ProcessingStep::actions);
+
+      nb::class_<FileInfo::NamedStats>(fi, "NamedStats", "a named SummaryStatistics block (-s)")
+          .def(nb::init<>())
+          .def_ro("title", &FileInfo::NamedStats::title)
+          .def_ro("stats", &FileInfo::NamedStats::stats);
+
+      nb::class_<FileInfo::PeakInfo>(fi, "PeakInfo", "peak-file (MSExperiment) specifics")
+          .def(nb::init<>())
+          .def_ro("instrument_name", &FileInfo::PeakInfo::instrument_name)
+          .def_ro("mass_analyzers", &FileInfo::PeakInfo::mass_analyzers)
+          .def_ro("ms_levels", &FileInfo::PeakInfo::ms_levels)
+          .def_ro("total_peaks", &FileInfo::PeakInfo::total_peaks)
+          .def_ro("num_spectra", &FileInfo::PeakInfo::num_spectra)
+          .def_ro("spectra_per_ms_level", &FileInfo::PeakInfo::spectra_per_ms_level)
+          .def_ro("peak_type_per_ms_level", &FileInfo::PeakInfo::peak_type_per_ms_level)
+          .def_ro("precursor_charges", &FileInfo::PeakInfo::precursor_charges)
+          .def_ro("float_arrays", &FileInfo::PeakInfo::float_arrays)
+          .def_ro("int_arrays", &FileInfo::PeakInfo::int_arrays)
+          .def_ro("string_arrays", &FileInfo::PeakInfo::string_arrays)
+          .def_ro("faims_cvs", &FileInfo::PeakInfo::faims_cvs)
+          .def_ro("num_chromatograms", &FileInfo::PeakInfo::num_chromatograms)
+          .def_ro("num_chrom_peaks", &FileInfo::PeakInfo::num_chrom_peaks)
+          .def_ro("chromatogram_types", &FileInfo::PeakInfo::chromatogram_types)
+          .def("activationMethodsFlat", [](const FileInfo::PeakInfo& self){ return self.activationMethodsFlat(); },
+               "Activation methods as a list of (ms_level, method_name, count) tuples");
+
+      nb::class_<FileInfo::FeatureInfo::MapColumn>(fi, "MapColumn", "a consensus map column header")
+          .def(nb::init<>())
+          .def_ro("filename", &FileInfo::FeatureInfo::MapColumn::filename)
+          .def_ro("identifier", &FileInfo::FeatureInfo::MapColumn::identifier)
+          .def_ro("label", &FileInfo::FeatureInfo::MapColumn::label)
+          .def_ro("size", &FileInfo::FeatureInfo::MapColumn::size);
+
+      nb::class_<FileInfo::FeatureInfo>(fi, "FeatureInfo", "feature / consensus specifics")
+          .def(nb::init<>())
+          .def_ro("is_consensus", &FileInfo::FeatureInfo::is_consensus)
+          .def_ro("num_features", &FileInfo::FeatureInfo::num_features)
+          .def_ro("tic", &FileInfo::FeatureInfo::tic)
+          .def_ro("charges", &FileInfo::FeatureInfo::charges)
+          .def_ro("ids_per_element", &FileInfo::FeatureInfo::ids_per_element)
+          .def_ro("assigned_ids", &FileInfo::FeatureInfo::assigned_ids)
+          .def_ro("unassigned_ids", &FileInfo::FeatureInfo::unassigned_ids)
+          .def_ro("size_distribution", &FileInfo::FeatureInfo::size_distribution)
+          .def_ro("map_columns", &FileInfo::FeatureInfo::map_columns);
+
+      nb::class_<FileInfo::IdentInfo>(fi, "IdentInfo", "identification (idXML / mzIdentML) specifics")
+          .def(nb::init<>())
+          .def_ro("db_name", &FileInfo::IdentInfo::db_name)
+          .def_ro("db_version", &FileInfo::IdentInfo::db_version)
+          .def_ro("taxonomy", &FileInfo::IdentInfo::taxonomy)
+          .def_ro("search_engines", &FileInfo::IdentInfo::search_engines)
+          .def_ro("num_runs", &FileInfo::IdentInfo::num_runs)
+          .def_ro("protein_hits", &FileInfo::IdentInfo::protein_hits)
+          .def_ro("non_redundant_protein_hits", &FileInfo::IdentInfo::non_redundant_protein_hits)
+          .def_ro("matched_spectra", &FileInfo::IdentInfo::matched_spectra)
+          .def_ro("peptide_hits", &FileInfo::IdentInfo::peptide_hits)
+          .def_ro("psms_per_spectrum", &FileInfo::IdentInfo::psms_per_spectrum)
+          .def_ro("avg_peptide_length", &FileInfo::IdentInfo::avg_peptide_length)
+          .def_ro("non_redundant_peptides", &FileInfo::IdentInfo::non_redundant_peptides)
+          .def_ro("modified_tophits", &FileInfo::IdentInfo::modified_tophits)
+          .def_ro("modification_counts", &FileInfo::IdentInfo::modification_counts);
+
+      nb::class_<FileInfo::FastaInfo>(fi, "FastaInfo", "FASTA specifics")
+          .def(nb::init<>())
+          .def_ro("num_sequences", &FileInfo::FastaInfo::num_sequences)
+          .def_ro("total_residues", &FileInfo::FastaInfo::total_residues)
+          .def_ro("is_nucleic_acid", &FileInfo::FastaInfo::is_nucleic_acid)
+          .def_ro("length_stats", &FileInfo::FastaInfo::length_stats)
+          .def_ro("seq_with_ambiguous", &FileInfo::FastaInfo::seq_with_ambiguous)
+          .def_ro("dup_headers", &FileInfo::FastaInfo::dup_headers)
+          .def_ro("dup_sequences", &FileInfo::FastaInfo::dup_sequences)
+          .def_ro("ambiguity_counts", &FileInfo::FastaInfo::ambiguity_counts);
+
+      nb::class_<FileInfo::MzTabInfo>(fi, "MzTabInfo", "mzTab specifics")
+          .def(nb::init<>())
+          .def_ro("version", &FileInfo::MzTabInfo::version)
+          .def_ro("mode", &FileInfo::MzTabInfo::mode)
+          .def_ro("type", &FileInfo::MzTabInfo::type)
+          .def_ro("psms", &FileInfo::MzTabInfo::psms)
+          .def_ro("peptides", &FileInfo::MzTabInfo::peptides)
+          .def_ro("proteins", &FileInfo::MzTabInfo::proteins)
+          .def_ro("oligonucleotides", &FileInfo::MzTabInfo::oligonucleotides)
+          .def_ro("osms", &FileInfo::MzTabInfo::osms)
+          .def_ro("small_molecules", &FileInfo::MzTabInfo::small_molecules)
+          .def_ro("nucleic_acids", &FileInfo::MzTabInfo::nucleic_acids);
+
+      nb::class_<FileInfo::ValidationInfo>(fi, "ValidationInfo", "the -v / -i blocks")
+          .def(nb::init<>())
+          .def_ro("performed", &FileInfo::ValidationInfo::performed)
+          .def_ro("supported", &FileInfo::ValidationInfo::supported)
+          .def_ro("valid", &FileInfo::ValidationInfo::valid)
+          .def_ro("schema_version", &FileInfo::ValidationInfo::schema_version)
+          .def_ro("detail", &FileInfo::ValidationInfo::detail)
+          .def_ro("warnings", &FileInfo::ValidationInfo::warnings)
+          .def_ro("errors", &FileInfo::ValidationInfo::errors)
+          .def_ro("index_checked", &FileInfo::ValidationInfo::index_checked)
+          .def_ro("index_valid", &FileInfo::ValidationInfo::index_valid)
+          .def_ro("indexed_spectra", &FileInfo::ValidationInfo::indexed_spectra)
+          .def_ro("indexed_chromatograms", &FileInfo::ValidationInfo::indexed_chromatograms);
+
+      nb::class_<FileInfo::CorruptionInfo>(fi, "CorruptionInfo", "the -c block")
+          .def(nb::init<>())
+          .def_ro("performed", &FileInfo::CorruptionInfo::performed)
+          .def_ro("errors", &FileInfo::CorruptionInfo::errors)
+          .def_ro("warnings", &FileInfo::CorruptionInfo::warnings);
+
+      nb::class_<FileInfo::DetailInfo>(fi, "DetailInfo", "the -d per-spectrum listing (pre-rendered)")
+          .def(nb::init<>())
+          .def_ro("performed", &FileInfo::DetailInfo::performed)
+          .def_ro("lines", &FileInfo::DetailInfo::lines);
+
+      nb::class_<FileInfo::Result>(fi, "Result", "all file-level information")
+          .def(nb::init<>())
+          .def_ro("meta", &FileInfo::Result::meta)
+          .def_ro("ranges", &FileInfo::Result::ranges)
+          .def_ro("peak", &FileInfo::Result::peak)
+          .def_ro("feature", &FileInfo::Result::feature)
+          .def_ro("ident", &FileInfo::Result::ident)
+          .def_ro("fasta", &FileInfo::Result::fasta)
+          .def_ro("mztab", &FileInfo::Result::mztab)
+          .def_ro("experiment_meta", &FileInfo::Result::experiment_meta)
+          .def_ro("processing", &FileInfo::Result::processing)
+          .def_ro("statistics", &FileInfo::Result::statistics)
+          .def_ro("validation", &FileInfo::Result::validation)
+          .def_ro("corruption", &FileInfo::Result::corruption)
+          .def_ro("detail", &FileInfo::Result::detail)
+          .def_ro("transformation_summary", &FileInfo::Result::transformation_summary)
+          .def_ro("targeted_summary", &FileInfo::Result::targeted_summary)
+          .def_ro("text", &FileInfo::Result::text)
+          .def_ro("tsv", &FileInfo::Result::tsv);
+
+      nb::class_<FileInfo::Options>(fi, "Options", "which optional analyses to run (mirror the CLI flags)")
+          .def(nb::init<>())
+          .def("__copy__", [](const FileInfo::Options& self){ return FileInfo::Options(self); })
+          .def("__deepcopy__", [](const FileInfo::Options& self, nb::dict){ return FileInfo::Options(self); }, "memo"_a)
+          .def_rw("forced_type", &FileInfo::Options::forced_type)
+          .def_rw("meta", &FileInfo::Options::meta)
+          .def_rw("processing", &FileInfo::Options::processing)
+          .def_rw("statistics", &FileInfo::Options::statistics)
+          .def_rw("detailed", &FileInfo::Options::detailed)
+          .def_rw("check_corrupt", &FileInfo::Options::check_corrupt)
+          .def_rw("validate", &FileInfo::Options::validate)
+          .def_rw("check_index", &FileInfo::Options::check_index);
+
+      fi.def(nb::init<>())
+        .def("__copy__", [](const FileInfo& self){ return FileInfo(self); })
+        .def("__deepcopy__", [](const FileInfo& self, nb::dict){ return FileInfo(self); }, "memo"_a)
+        .def("run", [](FileInfo& self, const std::string& filename, const FileInfo::Options& options){ return self.run(filename, options); },
+             "filename"_a, "options"_a, "Load the file (type auto-detected unless options.forced_type) and compute all requested information")
+        .def("run", [](FileInfo& self, const std::string& filename){ return self.run(filename); },
+             "filename"_a, "Load the file with default options")
+        .def("runAll", [](FileInfo& self, const std::string& filename){ return self.runAll(filename); },
+             "filename"_a, "Compute all content metrics (meta/processing/statistics on; no validation/index/detail/corrupt)")
+        .def_static("toText", [](const FileInfo::Result& r){ return FileInfo::toText(r); }, "result"_a,
+                    "Render the result as the FileInfo CLI human-readable text")
+        .def_static("toTSV", [](const FileInfo::Result& r){ return FileInfo::toTSV(r); }, "result"_a,
+                    "Render the result as the FileInfo CLI TSV");
+    }
 
 }
